# doc/model/swagger-2.0.yaml
swagger: '2.0'
info:
  description: REST API for the 5GCity's SDK
  version: '1.0'
  title: '5GCity SDK REST API '
  contact:
    name: Nextworks S.r.l.
    url: 'http://www.nextworks.it'
    email: info@nextworks.it
  license:
    name: Apache License Version 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0"'
host: 'localhost:8081'
basePath: /
...
paths:
...
  '/sdk/service_descriptor/{serviceDescriptorId}/publish':
    post:
      tags:
        - service-descriptors-controller
      summary: Publish SDK Service to Public Catalogue
      operationId: publishServiceUsingPOST_1
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: Authorization
          in: header
          description: Authorization
          required: false
          type: string
        - name: serviceDescriptorId
          in: path
          description: serviceDescriptorId
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: Created
        '202':
          description: Network Service descriptor created. The descriptor will be published to the Public Catalogue
          schema:
            type: object
        '400':
          description: Publish request without parameter serviceDescriptorId or SDK Service already published
        '401':
          description: User not authenticated
        '403':
          description: User not allowed to access the resource
        '404':
          description: SDK Service Descriptor not present in database
        '409':
          description: Not all components are published to the Public Catalogue
      deprecated: false
...
  /sdk/services/:
    get:
      tags:
        - service-controller
      summary: Get the complete list of SDK Services available in database
      operationId: getServicesUsingGET
      produces:
        - '*/*'
      parameters:
        - name: Authorization
          in: header
          description: Authorization
          required: false
          type: string
        - name: sliceId
          in: query
          description: sliceId
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/SdkService'
        '401':
          description: User not authenticated
        '403':
          description: User not allowed to access the resource
        '404':
          description: Slice not found
      deprecated: false
    post:
      tags:
        - service-controller
      summary: Create a new SDK Service
      operationId: createServiceUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: Authorization
          in: header
          description: Authorization
          required: false
          type: string
        - in: body
          name: request
          description: request
          required: true
          schema:
            $ref: '#/definitions/SdkService'
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: SDK Service created
          schema:
            type: object
        '400':
          description: SDK Service already present in database
        '401':
          description: User not authenticated
        '403':
          description: User not allowed to access the resource
        '404':
          description: SDK Service cannot be validated (some component is missing) or slice not found
      deprecated: false
    put:
      tags:
        - service-controller
      summary: Modify an existing SDK Service
      operationId: updateServiceUsingPUT
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: Authorization
          in: header
          description: Authorization
          required: false
          type: string
        - in: body
          name: request
          description: request
          required: true
          schema:
            $ref: '#/definitions/SdkService'
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: Created
        '204':
          description: SDK Service updated
          schema:
            type: object
        '400':
          description: SDK Service cannot be validated
        '401':
          description: User not authenticated
        '403':
          description: User not allowed to access the resource
        '404':
          description: SDK Service not present in database or slice not found
        '409':
          description: SDK Service cannot be updated
      deprecated: false
  '/sdk/services/service/{serviceId}/publish':
    post:
      tags:
        - service-controller
      summary: Publish SDK Service to Public Catalogue
      operationId: publishServiceUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: Authorization
          in: header
          description: Authorization
          required: false
          type: string
        - in: body
          name: makeDescriptorRequest
          description: makeDescriptorRequest
          required: true
          schema:
            $ref: '#/definitions/MakeDescriptorRequest'
        - name: serviceId
          in: path
          description: serviceId
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: Created
        '202':
          description: Network Service descriptor created. The descriptor will be published to the Public Catalogue
          schema:
            type: object
        '400':
          description: Publish request without parameter serviceId or provided parameters cannot be validated
        '401':
          description: User not authenticated
        '403':
          description: User not allowed to access the resource
        '404':
          description: SDK Service not present in database
        '409':
          description: Not all components are published to the Public Catalogue
      deprecated: false
  '/sdk/services/{serviceId}':
    get:
      tags:
        - service-controller
      summary: Search a SDK Service with ID
      operationId: getServiceUsingGET
      produces:
        - '*/*'
      parameters:
        - name: Authorization
          in: header
          description: Authorization
          required: false
          type: string
        - name: serviceId
          in: path
          description: serviceId
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/SdkService'
        '400':
          description: Query without parameter serviceId
        '401':
          description: User not authenticated
        '403':
          description: User not allowed to access the resource
        '404':
          description: SdkService not found in database
      deprecated: false
    delete:
      tags:
        - service-controller
      summary: Delete a SDK Service from database
      operationId: deleteServiceUsingDELETE
      produces:
        - '*/*'
      parameters:
        - name: Authorization
          in: header
          description: Authorization
          required: false
          type: string
        - name: serviceId
          in: path
          description: serviceId
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            type: object
        '204':
          description: SDK Service deleted
          schema:
            type: object
        '400':
          description: Deletion request without parameter serviceId
        '401':
          description: User not authenticated
        '403':
          description: User not allowed to access the resource
        '404':
          description: SDK Service not present in database
        '409':
          description: SDK Service cannot be deleted
      deprecated: false
  '/sdk/services/{serviceId}/create_descriptor':
    post:
      tags:
        - service-controller
      summary: Create descriptor for SDK Service
      operationId: createDescriptorUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: Authorization
          in: header
          description: Authorization
          required: false
          type: string
        - in: body
          name: makeDescriptorRequest
          description: makeDescriptorRequest
          required: true
          schema:
            $ref: '#/definitions/MakeDescriptorRequest'
        - name: serviceId
          in: path
          description: serviceId
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: SDK Service Descriptor created
          schema:
            type: object
        '400':
          description: Create descriptor request without serviceId or provided parameters cannot be validated
        '401':
          description: User not authenticated
        '403':
          description: User not allowed to access the resource
        '404':
          description: SDK Service not found
      deprecated: false
...
  '/sdk/services/{serviceId}/monitoring_params/{monitoringParameterId}':
    delete:
      tags:
        - service-controller
      summary: Delete monitoring param from SDK Service
      operationId: deleteMonitoringParametersForServiceUsingDELETE
      produces:
        - '*/*'
      parameters:
        - name: Authorization
          in: header
          description: Authorization
          required: false
          type: string
        - name: monitoringParameterId
          in: path
          description: monitoringParameterId
          required: true
          type: integer
          format: int64
        - name: serviceId
          in: path
          description: serviceId
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            type: object
        '204':
          description: Monitoring Parameter deleted
          schema:
            type: object
        '400':
          description: Deletion request without parameter serviceId or service cannot be validated
        '401':
          description: User not authenticated
        '403':
          description: User not allowed to access the resource
        '404':
          description: SDK Service or Monitoring Parameter not present in database
        '409':
          description: Monitoring Parameter cannot be deleted
      deprecated: false
...
definitions:
...
  MakeDescriptorRequest:
    type: object
    properties:
      parameterValues:
        type: array
        items:
          type: number
    title: MakeDescriptorRequest
...
  SdkService:
    type: object
    properties:
      accessLevel:
        type: integer
        format: int32
      actionRules:
        type: array
        items:
          $ref: '#/definitions/ServiceActionRule'
      actions:
        type: array
        items:
          $ref: '#/definitions/ServiceAction'
      component:
        type: array
        items:
          $ref: '#/definitions/SdkServiceComponent'
      connectionPoints:
        type: array
        items:
          $ref: '#/definitions/ConnectionPoint'
      designer:
        type: string
      extMonitoringParameters:
        type: array
        items:
          $ref: '#/definitions/MonitoringParameter'
      id:
        type: string
      intMonitoringParameters:
        type: array
        items:
          $ref: '#/definitions/MonitoringParameter'
      l3Connectivity:
        type: array
        items:
          $ref: '#/definitions/L3Connectivity'
      license:
        $ref: '#/definitions/License'
      link:
        type: array
        items:
          $ref: '#/definitions/Link'
      metadata:
        type: object
        additionalProperties:
          type: string
      name:
        type: string
      ownerId:
        type: string
      parameters:
        type: array
        items:
          type: string
      sliceId:
        type: string
      version:
        type: string
      visibility:
        type: string
        enum:
          - PUBLIC
          - PRIVATE
    title: SdkService
...
  ServiceAction:
    type: object
    properties:
      actionType:
        type: string
        enum:
          - SCALE_IN
          - SCALE_OUT
          - COMPONENT
      componentIndex:
        type: string
      id:
        type: string
      name:
        type: string
    title: ServiceAction
...